Move `RelativeDate` component to more general file
