+++ frontend/src/ui/Blocks/Series.tsx
@@ -1,5 +1,4 @@
 import React from "react";
-import { useTranslation } from "react-i18next";
 import { graphql, useFragment } from "react-relay";
 
 import { keyOfId } from "../../util";
@@ -11,6 +10,7 @@ import {
     SeriesBlockSeriesData$key,
 } from "../../query-types/SeriesBlockSeriesData.graphql";
 import { Thumbnail } from "../Video";
+import { RelativeDate } from "../time";
 
 
 type SharedProps = {
@@ -140,7 +140,7 @@ const GridTile: React.FC<GridTypeProps> = ({ event, realmPath, active }) => {
                     },
                 }}>{event.creator}</span>}
                 {/* `new Date` is well defined for our ISO Date strings */}
-                <CreationDate date={new Date(event.created)} />
+                <RelativeDate date={new Date(event.created)} />
             </div>
         </div>
     </>;
@@ -173,37 +173,5 @@ const GridTile: React.FC<GridTypeProps> = ({ event, realmPath, active }) => {
             to={`${realmPath}${realmPath.endsWith("/") ? "" : "/"}v/${keyOfId(event.id)}`}
             css={containerStyle}
         >{inner}</Link>;
-};
-
-type CreationDateProps = {
-    date: Date;
-};
-
-const CreationDate: React.FC<CreationDateProps> = ({ date }) => {
-    const { i18n } = useTranslation();
-    const secsAgo = Math.floor((Date.now() - date.getTime()) / 1000);
-
-    const prettyDate = (() => {
-        const intl = new Intl.RelativeTimeFormat(i18n.language);
-        if (secsAgo <= 55) {
-            return intl.format(-secsAgo, "second");
-        } else if (secsAgo <= 55 * 60) {
-            return intl.format(-Math.round(secsAgo / 60), "minute");
-        } else if (secsAgo <= 23 * 60 * 60) {
-            return intl.format(-Math.round(secsAgo / 60 / 60), "hour");
-        } else if (secsAgo <= 6 * 24 * 60 * 60) {
-            return intl.format(-Math.round(secsAgo / 24 / 60 / 60), "day");
-        } else if (secsAgo <= 3.5 * 7 * 24 * 60 * 60) {
-            return intl.format(-Math.round(secsAgo / 7 / 24 / 60 / 60), "week");
-        } else if (secsAgo <= 11 * 30.5 * 24 * 60 * 60) {
-            return intl.format(-Math.round(secsAgo / 30.5 / 24 / 60 / 60), "month");
-        } else {
-            return intl.format(-Math.round(secsAgo / 365.25 / 24 / 60 / 60), "year");
-        }
-    })();
-
-    const preciseDate = date.toLocaleString(i18n.language);
-
-    return <span title={preciseDate}>{prettyDate}</span>;
 };
